+++ test/unit/plugins/Plugins.spec.js
@@ -22,6 +22,7 @@ const pluginsQuantity = (method, quantity) => {
 };
 
 describe("Plugins", () => {
+  let callbacks;
   let sandbox;
   let coreMocks;
   let coreInstance;
@@ -32,6 +33,11 @@ describe("Plugins", () => {
 
   beforeEach(async () => {
     sandbox = sinon.createSandbox();
+    callbacks = {
+      addAlert: sandbox.stub(),
+      removeAlerts: sandbox.stub(),
+      renameAlerts: sandbox.stub(),
+    };
     sandbox.stub(tracer, "verbose");
     sandbox.stub(tracer, "debug");
     sandbox.stub(tracer, "error");
@@ -54,7 +60,7 @@ describe("Plugins", () => {
   describe("register method", () => {
     const METHOD = "Register";
     it("should do nothing if there are no plugins to register", async () => {
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 1))).toEqual(true);
     });
@@ -62,7 +68,7 @@ describe("Plugins", () => {
     it("should register object plugins", async () => {
       const fooPlugin = {};
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 2))).toEqual(true);
     });
@@ -72,19 +78,38 @@ describe("Plugins", () => {
         register: () => {},
       };
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 2))).toEqual(true);
     });
 
-    it("should register object plugins with register method passing to it the core itself", async () => {
+    it("should register object plugins with register method passing to them the core itself", async () => {
       const fooPlugin = {
         register: sinon.spy(),
       };
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       expect(fooPlugin.register.calledWith(coreInstance)).toEqual(true);
+    });
+
+    it("should register object plugins with register method passing to them custom methods", async () => {
+      expect.assertions(3);
+      const callToLoader = sandbox.stub();
+      loaderMocks.stubs.instance.new.returns(callToLoader);
+      const fooPlugin = {
+        register: (coreInstance, methods) => {
+          methods.loadMocks();
+          methods.addAlert("foo", "Foo message");
+          methods.removeAlerts();
+        },
+      };
+      configInstance.coreOptions.plugins = [fooPlugin];
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
+      await plugins.register();
+      expect(callToLoader.callCount).toEqual(1);
+      expect(callbacks.addAlert.calledWith("1:foo", "Foo message")).toEqual(true);
+      expect(callbacks.removeAlerts.calledWith("1:")).toEqual(true);
     });
 
     it("should not register object plugins with register method throwing an error", async () => {
@@ -94,7 +119,7 @@ describe("Plugins", () => {
         },
       };
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 1))).toEqual(true);
     });
@@ -102,18 +127,22 @@ describe("Plugins", () => {
     it("should not register strings as plugins", async () => {
       expect.assertions(2);
       configInstance.coreOptions.plugins = ["foo"];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
-      expect(console.log.calledWith("Error registering plugin")).toEqual(true);
+      expect(callbacks.addAlert.calledWith("register:1", 'Error registering plugin "1"')).toEqual(
+        true
+      );
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 1))).toEqual(true);
     });
 
     it("should not register booleans as plugins", async () => {
       expect.assertions(2);
       configInstance.coreOptions.plugins = [true];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
-      expect(console.log.calledWith("Error registering plugin")).toEqual(true);
+      expect(callbacks.addAlert.calledWith("register:1", 'Error registering plugin "1"')).toEqual(
+        true
+      );
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 1))).toEqual(true);
     });
 
@@ -121,11 +150,29 @@ describe("Plugins", () => {
       expect.assertions(3);
       const fooPlugin = sinon.spy();
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       expect(fooPlugin.calledWith(coreInstance)).toEqual(true);
       expect(fooPlugin.callCount).toEqual(1);
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 2))).toEqual(true);
+    });
+
+    it("should register function plugins executing them passing custom methods", async () => {
+      expect.assertions(3);
+      const callToLoader = sandbox.stub();
+      loaderMocks.stubs.instance.new.returns(callToLoader);
+      const fooPlugin = (coreInstance, methods) => {
+        methods.loadMocks();
+        methods.addAlert("foo", "Foo message");
+        methods.removeAlerts();
+        return {};
+      };
+      configInstance.coreOptions.plugins = [fooPlugin];
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
+      await plugins.register();
+      expect(callToLoader.callCount).toEqual(1);
+      expect(callbacks.addAlert.calledWith("1:foo", "Foo message")).toEqual(true);
+      expect(callbacks.removeAlerts.calledWith("1:")).toEqual(true);
     });
 
     it("should register function plugins returning a register method", async () => {
@@ -135,7 +182,7 @@ describe("Plugins", () => {
         register: spy,
       });
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       expect(spy.calledWith(coreInstance)).toEqual(true);
       expect(spy.callCount).toEqual(1);
@@ -149,7 +196,7 @@ describe("Plugins", () => {
         },
       });
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 1))).toEqual(true);
     });
@@ -166,11 +213,31 @@ describe("Plugins", () => {
         }
       }
       configInstance.coreOptions.plugins = [FooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       expect(receivedCore).toEqual(coreInstance);
       expect(instantiated).toEqual(true);
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 2))).toEqual(true);
+    });
+
+    it("should register class plugins, instantiating them passing custom methods", async () => {
+      expect.assertions(3);
+      const callToLoader = sandbox.stub();
+      loaderMocks.stubs.instance.new.returns(callToLoader);
+      class FooPlugin {
+        constructor(core, methods) {
+          methods.loadMocks();
+          methods.addAlert("foo", "Foo message");
+          methods.removeAlerts();
+          return {};
+        }
+      }
+      configInstance.coreOptions.plugins = [FooPlugin];
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
+      await plugins.register();
+      expect(callToLoader.callCount).toEqual(1);
+      expect(callbacks.addAlert.calledWith("1:foo", "Foo message")).toEqual(true);
+      expect(callbacks.removeAlerts.calledWith("1:")).toEqual(true);
     });
 
     it("should not register class plugins if class throw an error when being created", async () => {
@@ -181,9 +248,11 @@ describe("Plugins", () => {
         }
       }
       configInstance.coreOptions.plugins = [FooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
-      expect(console.log.calledWith("Error registering plugin")).toEqual(true);
+      expect(callbacks.addAlert.calledWith("register:1", 'Error registering plugin "1"')).toEqual(
+        true
+      );
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 1))).toEqual(true);
     });
 
@@ -201,7 +270,7 @@ describe("Plugins", () => {
         }
       }
       configInstance.coreOptions.plugins = [FooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       expect(receivedCore).toEqual(coreInstance);
       expect(instantiated).toEqual(true);
@@ -219,13 +288,13 @@ describe("Plugins", () => {
         }
       }
       configInstance.coreOptions.plugins = [FooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 1))).toEqual(true);
     });
 
     it("should trace the total number of registered plugins", async () => {
-      expect.assertions(2);
+      expect.assertions(4);
       class FooPlugin {
         constructor() {
           throw new Error();
@@ -241,9 +310,17 @@ describe("Plugins", () => {
         "foo",
         { foo: "foo" },
       ];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
-      expect(console.log.calledWith("Error registering plugin")).toEqual(true);
+      expect(callbacks.addAlert.calledWith("register:4", 'Error registering plugin "4"')).toEqual(
+        true
+      );
+      expect(callbacks.addAlert.calledWith("register:5", 'Error registering plugin "5"')).toEqual(
+        true
+      );
+      expect(callbacks.addAlert.calledWith("register:6", 'Error registering plugin "6"')).toEqual(
+        true
+      );
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 4))).toEqual(true);
     });
   });
@@ -252,7 +329,7 @@ describe("Plugins", () => {
     const METHOD = "Initializat";
     it("should do nothing if there are no plugins to register", async () => {
       configInstance.coreOptions.plugins = null;
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.init();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 0))).toEqual(true);
@@ -264,7 +341,7 @@ describe("Plugins", () => {
         init: sinon.spy(),
       };
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.init();
       expect(fooPlugin.init.callCount).toEqual(1);
@@ -277,7 +354,7 @@ describe("Plugins", () => {
         displayName: "foo-plugin",
       };
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.init();
       expect(tracer.debug.calledWith('Initializing plugin "foo-plugin"')).toEqual(true);
@@ -292,14 +369,14 @@ describe("Plugins", () => {
         init: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.init();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 2))).toEqual(true);
     });
 
-    it("should catch init method errors", async () => {
-      expect.assertions(1);
+    it("should catch init method errors and notify alerts", async () => {
+      expect.assertions(3);
       const fooPlugin = {
         init: () => {
           throw new Error();
@@ -312,9 +389,13 @@ describe("Plugins", () => {
         init: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2, fooPlugin3];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.init();
+      expect(callbacks.removeAlerts.calledWith("init")).toEqual(true);
+      expect(callbacks.addAlert.calledWith("init:1", 'Error initializating plugin "1"')).toEqual(
+        true
+      );
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 2))).toEqual(true);
     });
 
@@ -334,7 +415,7 @@ describe("Plugins", () => {
         init: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2, fooPlugin3];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.init();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 2))).toEqual(true);
@@ -350,7 +431,7 @@ describe("Plugins", () => {
         init: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2, fooPlugin3];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.init();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 2))).toEqual(true);
@@ -362,7 +443,7 @@ describe("Plugins", () => {
 
     it("should do nothing if there are no plugins to register", async () => {
       configInstance.coreOptions.plugins = null;
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.start();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 1))).toEqual(true);
@@ -374,7 +455,7 @@ describe("Plugins", () => {
         start: sinon.spy(),
       };
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.start();
       expect(fooPlugin.start.callCount).toEqual(1);
@@ -387,7 +468,7 @@ describe("Plugins", () => {
         displayName: "foo-plugin",
       };
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.start();
       expect(tracer.debug.calledWith('Starting plugin "foo-plugin"')).toEqual(true);
@@ -402,14 +483,14 @@ describe("Plugins", () => {
         start: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.start();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 3))).toEqual(true);
     });
 
-    it("should catch start method errors", async () => {
-      expect.assertions(1);
+    it("should catch start method errors and notify alert", async () => {
+      expect.assertions(3);
       const fooPlugin = {
         start: () => {
           throw new Error();
@@ -422,9 +503,11 @@ describe("Plugins", () => {
         start: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2, fooPlugin3];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.start();
+      expect(callbacks.removeAlerts.calledWith("start")).toEqual(true);
+      expect(callbacks.addAlert.calledWith("start:1", 'Error starting plugin "1"')).toEqual(true);
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 3))).toEqual(true);
     });
 
@@ -444,7 +527,7 @@ describe("Plugins", () => {
         start: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2, fooPlugin3];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.start();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 3))).toEqual(true);
@@ -460,7 +543,7 @@ describe("Plugins", () => {
         start: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2, fooPlugin3];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.start();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 3))).toEqual(true);
@@ -472,7 +555,7 @@ describe("Plugins", () => {
 
     it("should do nothing if there are no plugins to stop", async () => {
       configInstance.coreOptions.plugins = null;
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.stop();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 1))).toEqual(true);
@@ -484,7 +567,7 @@ describe("Plugins", () => {
         stop: sinon.spy(),
       };
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.stop();
       expect(fooPlugin.stop.callCount).toEqual(1);
@@ -497,7 +580,7 @@ describe("Plugins", () => {
         displayName: "foo-plugin",
       };
       configInstance.coreOptions.plugins = [fooPlugin];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.stop();
       expect(tracer.debug.calledWith('Stopping plugin "foo-plugin"')).toEqual(true);
@@ -512,14 +595,14 @@ describe("Plugins", () => {
         stop: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.stop();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 3))).toEqual(true);
     });
 
-    it("should catch stop method errors", async () => {
-      expect.assertions(1);
+    it("should catch stop method errors and notify alert", async () => {
+      expect.assertions(3);
       const fooPlugin = {
         stop: () => {
           throw new Error();
@@ -532,9 +615,11 @@ describe("Plugins", () => {
         stop: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2, fooPlugin3];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.stop();
+      expect(callbacks.removeAlerts.calledWith("stop")).toEqual(true);
+      expect(callbacks.addAlert.calledWith("stop:1", 'Error stopping plugin "1"')).toEqual(true);
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 3))).toEqual(true);
     });
 
@@ -554,7 +639,7 @@ describe("Plugins", () => {
         stop: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2, fooPlugin3];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.stop();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 3))).toEqual(true);
@@ -570,10 +655,44 @@ describe("Plugins", () => {
         stop: () => Promise.resolve(),
       };
       configInstance.coreOptions.plugins = [fooPlugin, fooPlugin2, fooPlugin3];
-      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance);
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
       await plugins.register();
       await plugins.stop();
       expect(tracer.verbose.calledWith(pluginsQuantity(METHOD, 3))).toEqual(true);
+    });
+  });
+
+  describe("alerts", () => {
+    let fooPlugin;
+    beforeEach(() => {
+      fooPlugin = {
+        register: (core, { addAlert }) => {
+          addAlert("test-register", "Testing register alert");
+        },
+        start: (core, { addAlert }) => {
+          addAlert("test-start", "Testing start alert");
+        },
+        displayName: "foo-name",
+      };
+      configInstance.coreOptions.plugins = [fooPlugin];
+      plugins = new Plugins(configInstance, loaderMocks.stubs.instance, coreInstance, callbacks);
+    });
+
+    it("should have as scope the plugin index during the register method", async () => {
+      await plugins.register();
+      expect(callbacks.addAlert.calledWith("1:test-register", "Testing register alert")).toEqual(
+        true
+      );
+    });
+
+    it("should rename the scope if an alert is added during the start method", async () => {
+      expect.assertions(2);
+      await plugins.register();
+      await plugins.start();
+      expect(callbacks.renameAlerts.calledWith("1:", "foo-name:")).toEqual(true);
+      expect(callbacks.addAlert.calledWith("foo-name:test-start", "Testing start alert")).toEqual(
+        true
+      );
     });
   });
 });