+++ CHANGELOG.md
@@ -19,11 +19,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Remove behavior "name" property. Use id instead.
 - Remove behaviors "currentName" getter. Use "currentId" instead
 - Remove behaviors "names" getter. Use "ids" instead
+- Remove "serverError" getter. Use alerts instead
 
 ## [unreleased]
 ### Added
+- feat: Add alerts handler, allowing plugins and other internal pieces to inform about errors or warnings, and let plugins access to the list, so they can display it, for example. (#102)
+- chore(lint): Add eslint plugin to avoid the usage of `only`.
 ### Changed
+- feat: Rename filesLoader plugin
+- test: Move e2e helpers to a subfolder
 ### Fixed
+- fix: Avoid errors when calling to fixtures and behaviors getters if they are not already initialized
 ### Removed
 
 ## [1.5.2] - 2020-12-21